Move ReferencingNotebooks to sharable file
