+++ packages/components/src/entities/SampleTypeAppContext.tsx
@@ -15,8 +15,9 @@ import {
 } from '../internal/components/samples/models';
 import { DetailRenderer } from '../internal/components/forms/detail/DetailDisplay';
 
-import { ReferencingNotebooks, SampleGridButton, SampleStorageLocation, SampleStorageMenu } from './models';
+import { SampleGridButton, SampleStorageLocation, SampleStorageMenu } from './models';
 import { SamplesEditableGridProps } from './SamplesEditableGrid';
+import { ReferencingNotebooks } from '../internal/app/models';
 
 export interface SampleTypeAppContext {
     AddSamplesToStorageModalComponent: AddSamplesToStorageModal;
